vps/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
yt-dlp==2025.6.9
ffmpeg-python==0.2.0
python-multipart==0.0.6